fix README and enhance regression tests
README.md in template.tmpl had an error which wasn't caught.  template-helper
now exits with an error when gomplate exits with an error.  find also didn't exit
with an error when using -exec because find works that way.  changed to using xargs
for the same purpose as -exec which does (portably) exit with an error code when the
underlying script fails.

also fixed an error where the test would fail because of the new gpg signing config.
Makefile now checks if there's a gnupg folder in the homedir and skips configuring
signing if one isn't found (and prints a message to say so)